chore(textlint): use shallowMerge
